pyproject.toml
[project]
name = "todoproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn==0.27.0",
    "sqlalchemy==2.0.25",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "pydantic==2.6.1",
    "pydantic-settings==2.2.1",
    "email-validator==2.3.0",
    "python-jose==3.3.0",
    "pwdlib[argon2,bcrypt]==0.3.0",
    "PyJWT==2.11.0",
    "python-multipart==0.0.9",
]

[project.optional-dependencies]
dev = [
    "httpx==0.26.0"
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pylint>=4.0.5",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.8",
]
